updated Kubernetes_cluser_info.sh

+++ Kubernetes_cluser_info.sh
@@ -9,6 +9,7 @@ show_menu(){
     printf "${menu}**${number} 1)${menu} Show running pods and namespace ${normal}\n"
     printf "${menu}**${number} 2)${menu} Show creation time details of pods  ${normal}\n"
     printf "${menu}**${number} 3)${menu} Tell count of number of pods running ${normal}\n"
+    printf "${menu}**${number} 4)${menu} Check if your pod is running ${normal}\n"
     printf "${menu}*********************************************${normal}\n"
     printf "Please enter a menu option and enter OR ${fgred}x to exit from menu. ${normal}"
     read opt
@@ -45,9 +46,31 @@ while [ $opt != '' ]
             kubectl get pods --all-namespaces -o json | jq '.items[] | .spec.nodeName' -r | sort | uniq -c > out
             count=`cut -f 6 -d " " out`
             echo "Hey , count of pods running in cluster is $count "
-            rm -rf out 
+            rm -rf out; 
             show_menu;
         ;;
+	4) clear;
+	   option_picked "Option 4 Picked";
+	   printf "Please enter name of your pod --> ${normal}"
+           read opt 
+	   #printf "Your pod name is $opt \n"
+	   kubectl get pods --all-namespaces > out
+	   grep -i $opt out > out1
+	   grep -i running out1
+	   isInFile=$(cat out1 | grep -ic running)
+	   if [ $isInFile -eq 0 ]; then
+          	 printf "\n${fgred}Oops !! POD $opt is not running,Please check !! \n"
+	   else
+          	 printf "\n${number}Woww !! Your POD $opt is running \n"
+	  	 #kubectl describe pod $opt > out2
+	  	 #cat out2|grep -i namespace|rev| cut -d':' -f 1 | rev > out3
+                 namespace=$(cat out1 | cut -f1 -d" ")
+	  	 echo "Namespace for POD $opt is -> $namespace"
+          	 rm -rf out1;
+		 rm -rf out;
+           fi           
+           show_menu;
+	;;
         x)exit;
         ;;
         \n)exit;
